+++ database/migrations/2016_09_30_214651_cascade_delete_note_tags.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CascadeDeleteNoteTags extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('note_tag', function ($table) {
+            $table->dropForeign('note_tag_note_id_foreign');
+            $table->foreign('note_id')->references('id')->on('notes')->onDelete('cascade');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        //
+    }
+}